Defined static methods and renamed file to utils.py
