+++ estimators/test/utils.py
@@ -1,9 +1,10 @@
 import random, copy
 
 class Helper():
-    ''' Helper Function for tests '''
+    ''' Helper Class for tests '''
 
-    def run_add_example(self, datagen, class_object, num_examples):
+    @staticmethod
+    def run_add_example(datagen, class_object, num_examples):
         # class_object is the object of class Estimator() or class Interval()
         Object = copy.deepcopy(class_object)
 
@@ -17,19 +18,21 @@ class Helper():
 
         return Object
 
-    def get_estimate(self, datagen, listofestimators, num_examples):
+    @staticmethod
+    def get_estimate(datagen, listofestimators, num_examples):
         estimates = []
         for Estimator in listofestimators:
             # Estimator is a tuple
             # Estimator[0] is object of class Estimator()
             # Estimator[1] is expected value of the estimator
             
-            estimator = self.run_add_example(datagen, Estimator[0], num_examples)
+            estimator = Helper.run_add_example(datagen, Estimator[0], num_examples)
             estimates.append(estimator.get())
 
         return estimates
 
-    def calc_CI_width(self, datagen, listofintervals, n1, n2):
+    @staticmethod
+    def calc_CI_width(datagen, listofintervals, n1, n2):
         ''' Note: n1 is smaller than n2 '''
 
         width_n1 = []
@@ -37,12 +40,12 @@ class Helper():
         for interval in listofintervals:
 
             # For n1 number of examples
-            interval_n1 = self.run_add_example(datagen, interval, n1)
+            interval_n1 = Helper.run_add_example(datagen, interval, n1)
             result_n1 = interval_n1.get()
             width_n1.append(result_n1[1]-result_n1[0])
 
             # For n2 number of examples
-            interval_n2 = self.run_add_example(datagen, interval, n2)
+            interval_n2 = Helper.run_add_example(datagen, interval, n2)
             result_n2 = interval_n2.get()
             width_n2.append(result_n2[1]-result_n2[0])
